+++ examples/children/src/App.tsx
@@ -2,11 +2,15 @@ import * as React from "react";
 
 import {
   defineRouterConfig,
-  MRouter as Router,
+  MRouter,
+  MHRouter,
   Link,
   Outlet,
   useRouter
 } from "react-router-manage";
+
+const Router = window.__INITIAL_DATA__.mode === "hash" ? MHRouter : MRouter;
+
 
 function Layout({ children }) {
   return (
@@ -114,6 +118,7 @@ function Children() {
 }
 
 const routerConfig = defineRouterConfig({
+  basename: window.__INITIAL_DATA__.mode !== 'hash' ? window.__INITIAL_DATA__.basename : '/',
   routes: [
     {
       path: "/",