+++ examples/lazy-loading/src/vite-env.d.ts
@@ -1,1 +1,7 @@
 /// <reference types="vite/client" />
+interface Window {
+    __INITIAL_DATA__: {
+        basename: string;
+        mode: string;
+    }
+}